Refactored folder structure
