+++ cli/app/src/main/java/org/ammonium/linkit/cli/Add.java
@@ -1,4 +1,4 @@
-package org.ammonium.linkit.cli.add;
+package org.ammonium.linkit.cli;
 
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
